requirements-dev.txt
-r requirements.txt
django-debug-toolbar==1.9.1
codecov==2.0.15
coverage==4.5.1
selenium==3.12.0
sqlparse==0.2.4
